+++ functions.php
@@ -33,6 +33,9 @@ $bootscore_includes = array(
 );
 
 
+
+
+require_once('inc/breadcrumb.php');
 require_once('inc/widgets.php');
 
 
@@ -148,150 +151,8 @@ function bootscore_content_width() {
 add_action('after_setup_theme', 'bootscore_content_width', 0);
 
 
-/**
- * Register widget area.
- *
- * @link https://developer.wordpress.org/themes/functionality/sidebars/#registering-a-sidebar
- */
-/*
-// Widgets
-if (!function_exists('bootscore_widgets_init')) :
 
-  function bootscore_widgets_init() {
 
-    // Top Nav
-    register_sidebar(array(
-      'name'          => esc_html__('Top Nav', 'bootscore'),
-      'id'            => 'top-nav',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="top-nav-widget ms-1 ms-md-2">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<div class="widget-title d-none">',
-      'after_title'   => '</div>'
-    ));
-
-    // Top Nav 2
-    // Adds a widget next to the Top Nav position but moves to offcanvas on <lg breakpoint
-    register_sidebar(array(
-      'name'          => esc_html__('Top Nav 2', 'bootscore'),
-      'id'            => 'top-nav-2',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="top-nav-widget-2 d-lg-flex align-items-lg-center mt-2 mt-lg-0 ms-lg-2">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<div class="widget-title d-none">',
-      'after_title'   => '</div>'
-    ));
-
-    // Top Nav Search
-    register_sidebar(array(
-      'name'          => esc_html__('Top Nav Search', 'bootscore'),
-      'id'            => 'top-nav-search',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="top-nav-search">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<div class="widget-title d-none">',
-      'after_title'   => '</div>'
-    ));
-
-    // Sidebar
-    register_sidebar(array(
-      'name'          => esc_html__('Sidebar', 'bootscore'),
-      'id'            => 'sidebar-1',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<section id="%1$s" class="widget %2$s card card-body mb-4">',
-      'after_widget'  => '</section>',
-      'before_title'  => '<h2 class="widget-title card-header h5">',
-      'after_title'   => '</h2>',
-    ));
-
-    // Top Footer
-    register_sidebar(array(
-      'name'          => esc_html__('Top Footer', 'bootscore'),
-      'id'            => 'top-footer',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="footer_widget mb-5">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<h2 class="widget-title">',
-      'after_title'   => '</h2>'
-    ));
-
-    // Footer 1
-    register_sidebar(array(
-      'name'          => esc_html__('Footer 1', 'bootscore'),
-      'id'            => 'footer-1',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="footer_widget mb-4">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<h2 class="widget-title h5">',
-      'after_title'   => '</h2>'
-    ));
-
-    // Footer 2
-    register_sidebar(array(
-      'name'          => esc_html__('Footer 2', 'bootscore'),
-      'id'            => 'footer-2',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="footer_widget mb-4">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<h2 class="widget-title h5">',
-      'after_title'   => '</h2>'
-    ));
-
-    // Footer 3
-    register_sidebar(array(
-      'name'          => esc_html__('Footer 3', 'bootscore'),
-      'id'            => 'footer-3',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="footer_widget mb-4">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<h2 class="widget-title h5">',
-      'after_title'   => '</h2>'
-    ));
-
-    // Footer 4
-    register_sidebar(array(
-      'name'          => esc_html__('Footer 4', 'bootscore'),
-      'id'            => 'footer-4',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="footer_widget mb-4">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<h2 class="widget-title h5">',
-      'after_title'   => '</h2>'
-    ));
-
-    // Footer Info
-    register_sidebar(array(
-      'name'          => esc_html__('Footer Info', 'bootscore'),
-      'id'            => 'footer-info',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="footer_widget">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<div class="widget-title d-none">',
-      'after_title'   => '</div>'
-    ));
-
-    // 404 Page
-    register_sidebar(array(
-      'name'          => esc_html__('404 Page', 'bootscore'),
-      'id'            => '404-page',
-      'description'   => esc_html__('Add widgets here.', 'bootscore'),
-      'before_widget' => '<div class="mb-4">',
-      'after_widget'  => '</div>',
-      'before_title'  => '<h1 class="widget-title">',
-      'after_title'   => '</h1>'
-    ));
-
-  }
-
-  add_action('widgets_init', 'bootscore_widgets_init');
-
-endif;
-// Widgets END
-*/
-
-// Shortcode in HTML-Widget
-add_filter('widget_text', 'do_shortcode');
-// Shortcode in HTML-Widget End
 
 
 //Enqueue scripts and styles
@@ -462,34 +323,7 @@ add_post_type_support('page', 'excerpt');
 // Excerpt to pages END
 
 
-// Breadcrumb
-if (!function_exists('the_breadcrumb')) :
-  function the_breadcrumb() {
 
-    if (!is_home()) {
-      echo '<nav aria-label="breadcrumb" class="breadcrumb-scroller mb-4 mt-2 py-2 px-3 bg-body-tertiary rounded">';
-      echo '<ol class="breadcrumb mb-0">';
-      echo '<li class="breadcrumb-item"><a href="' . home_url() . '">' . '<i class="fa-solid fa-house"></i>' . '</a></li>';
-      // display parent category names with links
-      if (is_category() || is_single()) {
-        $cat_IDs = wp_get_post_categories(get_the_ID());
-        foreach ($cat_IDs as $cat_ID) {
-          $cat = get_category($cat_ID);
-          echo '<li class="breadcrumb-item"><a href="' . get_term_link($cat->term_id) . '">' . $cat->name . '</a></li>';
-        }
-      }
-      // display current page name
-      if (is_page() || is_single()) {
-        echo '<li class="breadcrumb-item active" aria-current="page">' . get_the_title() . '</li>';
-      }
-      echo '</ol>';
-      echo '</nav>';
-    }
-  }
-
-  add_filter('breadcrumbs', 'breadcrumbs');
-endif;
-// Breadcrumb END
 
 
 // Comment Button
